requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-dotenv==1.0.1
httpx==0.27.2
pydantic>=2.11.0,<3.0.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
gunicorn==23.0.0
pytest==8.3.3
pytest-cov==6.0.0
